examples/subobc: apply rename migration(TlmCmd -> tlm_cmd)
