+++ examples/subobc/src/src_user/Drivers/Etc/mobc.c
@@ -7,7 +7,7 @@
 // FIXME: DS 側の TCP が整理されたら， TCP 関連を撲滅し， CTCP に統一する
 
 #include "./mobc.h"
-#include <src_core/TlmCmd/common_tlm_cmd_packet.h>
+#include <src_core/tlm_cmd/common_tlm_cmd_packet.h>
 #include <src_core/Drivers/Protocol/eb90_frame_for_driver_super.h>
 #include <src_core/Drivers/Protocol/common_tlm_cmd_packet_for_driver_super.h>
 #include <string.h>
