+++ examples/subobc/src/src_user/Drivers/Etc/mobc.h
@@ -8,8 +8,8 @@
 #include <src_core/hal/uart.h>
 #include <src_core/Drivers/Super/driver_super.h>
 #include <src_core/system/time_manager/obc_time.h>
-#include <src_core/TlmCmd/common_tlm_packet.h>
-#include <src_core/TlmCmd/packet_handler.h>
+#include <src_core/tlm_cmd/common_tlm_packet.h>
+#include <src_core/tlm_cmd/packet_handler.h>
 
 /**
  * @enum   MOBC_TX_ERR_CODE
