+++ examples/subobc/src/src_user/Settings/Modes/TaskLists/Elements/tl_elem_gs_related_process.c
@@ -1,10 +1,10 @@
 #pragma section REPRO
 #include "tl_elem_gs_related_process.h"
 
-#include "../../../../TlmCmd/block_command_definitions.h"
+#include "../../../../tlm_cmd/block_command_definitions.h"
 #include "../../../../applications/app_registry.h"
 
-#include <src_core/TlmCmd/block_command_loader.h>
+#include <src_core/tlm_cmd/block_command_loader.h>
 
 void BCL_load_tl_elem_gs_related_process(void)
 {
