+++ examples/subobc/src/src_user/Settings/tlm_cmd/common_cmd_packet_define.h
@@ -5,8 +5,8 @@
 #ifndef COMMON_CMD_PACKET_DEFINE_H_
 #define COMMON_CMD_PACKET_DEFINE_H_
 
-#include <src_core/TlmCmd/Ccsds/space_packet_typedef.h>
-#include "./Ccsds/apid_define.h"
+#include <src_core/tlm_cmd/ccsds/space_packet_typedef.h>
+#include "./ccsds/apid_define.h"
 
 // CommonCmdPacket として CmdSpacePacket をつかう
 typedef CmdSpacePacket CommonCmdPacket;
