+++ examples/subobc/src/src_user/Settings/tlm_cmd/common_tlm_cmd_packet_define.h
@@ -5,7 +5,7 @@
 #ifndef COMMON_TLM_CMD_PACKET_DEFINE_H_
 #define COMMON_TLM_CMD_PACKET_DEFINE_H_
 
-#include <src_core/TlmCmd/Ccsds/space_packet_typedef.h>
+#include <src_core/tlm_cmd/ccsds/space_packet_typedef.h>
 
 // CommonTlmCmdPacket として SpacePacket をつかう
 typedef SpacePacket CommonTlmCmdPacket;
