+++ examples/subobc/src/src_user/Settings/tlm_cmd/data_base/CMD_DB/SAMPLE_AOBC_CMD_DB.xlsm
