+++ examples/subobc/src/src_user/applications/driver_instances/di_mobc.h
@@ -7,7 +7,7 @@
 
 #include "../../Drivers/Etc/mobc.h"
 #include <src_core/system/application_manager/app_info.h>
-#include <src_core/TlmCmd/common_cmd_packet.h>
+#include <src_core/tlm_cmd/common_cmd_packet.h>
 
 extern const MOBC_Driver* const mobc_driver;           //!< MOBC driver
