+++ examples/subobc/src/src_user/tlm_cmd/block_command_definitions.h
@@ -27,7 +27,7 @@ typedef enum
   BC_AC_TLM_CMD_HIRATE = 50,
 
   // ==== 各系領域 ====
-  // ./C2A/TlmCmd/NormalBlockCommandDefinition/で定義
+  // ./C2A/tlm_cmd/normal_block_command_definition/で定義
   // イベントハンドラはBC_EH_など，接頭辞を適切につけること！
 
   // CDH:60-79
