+++ examples/subobc/src/src_user/tlm_cmd/command_definitions.c
@@ -4,7 +4,7 @@
  * @brief  コマンド定義
  * @note   このコードは自動生成されています！
  */
-#include <src_core/TlmCmd/command_analyze.h>
+#include <src_core/tlm_cmd/command_analyze.h>
 #include "command_definitions.h"
 #include "command_source.h"
