+++ examples/subobc/src/src_user/tlm_cmd/common_cmd_packet.c
@@ -4,8 +4,8 @@
  * @brief  CTP 関連基本関数の定義
  * @note   CCP:  CommonCmdPacket
  */
-#include <src_core/TlmCmd/common_cmd_packet.h>
-#include <src_core/TlmCmd/Ccsds/cmd_space_packet.h>
+#include <src_core/tlm_cmd/common_cmd_packet.h>
+#include <src_core/tlm_cmd/ccsds/cmd_space_packet.h>
 #include <string.h>
