+++ examples/subobc/src/src_user/tlm_cmd/common_tlm_cmd_packet.c
@@ -6,8 +6,8 @@
  * @note   CTP:  CommonTlmPacket
  * @note   CCP:  CommonCmdPacket
  */
-#include <src_core/TlmCmd/common_tlm_cmd_packet.h>
-#include <src_core/TlmCmd/Ccsds/space_packet.h>
+#include <src_core/tlm_cmd/common_tlm_cmd_packet.h>
+#include <src_core/tlm_cmd/ccsds/space_packet.h>
 #include <string.h>
