+++ examples/subobc/src/src_user/tlm_cmd/common_tlm_packet.c
@@ -4,8 +4,8 @@
  * @brief  CTP 関連基本関数の定義
  * @note   CTP:  CommonTlmPacket
  */
-#include <src_core/TlmCmd/common_tlm_packet.h>
-#include <src_core/TlmCmd/Ccsds/tlm_space_packet.h>
+#include <src_core/tlm_cmd/common_tlm_packet.h>
+#include <src_core/tlm_cmd/ccsds/tlm_space_packet.h>
 #include <string.h>
