+++ examples/subobc/src/src_user/tlm_cmd/normal_block_command_definition/nbc_header.h
@@ -1,7 +1,7 @@
 #ifndef NBC_HEADER_H_
 #define NBC_HEADER_H_
 
-#include <src_core/TlmCmd/block_command_table.h>
+#include <src_core/tlm_cmd/block_command_table.h>
 
 void BCL_load_start_hk_tlm(void);
