+++ examples/subobc/src/src_user/tlm_cmd/normal_block_command_definition/nbc_start_hk_tlm.c
@@ -6,7 +6,7 @@
 #include "../telemetry_definitions.h"
 
 #include <src_core/applications/timeline_command_dispatcher_id_define.h>
-#include <src_core/TlmCmd/block_command_loader.h>
+#include <src_core/tlm_cmd/block_command_loader.h>
 
 
 void BCL_load_start_hk_tlm(void)
