+++ examples/subobc/src/src_user/tlm_cmd/telemetry_definitions.c
@@ -4,7 +4,7 @@
  * @brief  テレメトリ定義
  * @note   このコードは自動生成されています！
  */
-#include <src_core/TlmCmd/telemetry_frame.h>
+#include <src_core/tlm_cmd/telemetry_frame.h>
 #include "telemetry_definitions.h"
 #include "telemetry_source.h"
