+++ examples/subobc/sync_with_mobc_example.bat
@@ -11,7 +11,7 @@ call :sync_dir ".\src\src_user\BootLoader\" "..\mobc\src\src_user\BootLoader\"
 call :sync_dir ".\src\src_user\library\" "..\mobc\src\src_user\library\"
 call :sync_dir ".\src\src_user\ReProgramming\" "..\mobc\src\src_user\ReProgramming\"
 call :sync_dir ".\src\src_user\script\" "..\mobc\src\src_user\script\"
-call :sync_dir ".\src\src_user\Settings\TlmCmd\Ccsds\" "..\mobc\src\src_user\Settings\TlmCmd\Ccsds\"
+call :sync_dir ".\src\src_user\Settings\tlm_cmd\ccsds\" "..\mobc\src\src_user\Settings\tlm_cmd\ccsds\"
 
 call :sync_file ".\src\src_user\applications\user_defined\debug_apps.h" "..\mobc\src\src_user\applications\user_defined\debug_apps.h"
 call :sync_file ".\src\src_user\hal\sils\uart_sils_sci_if.hpp" "..\mobc\src\src_user\hal\sils\uart_sils_sci_if.hpp"
@@ -21,13 +21,13 @@ call :sync_file ".\src\src_user\hal\sils_mockup\uart_sils.c" "..\mobc\src\src_us
 call :sync_file ".\src\src_user\hal\sils_mockup\wdt_sils.c" "..\mobc\src\src_user\hal\sils_mockup\wdt_sils.c"
 call :sync_file ".\src\src_user\Settings\git_revision_config.h" "..\mobc\src\src_user\Settings\git_revision_config.h"
 call :sync_file ".\src\src_user\Settings\DriverSuper\driver_buffer_define.h" "..\mobc\src\src_user\Settings\DriverSuper\driver_buffer_define.h"
-call :sync_file ".\src\src_user\Settings\TlmCmd\common_cmd_packet_define.c" "..\mobc\src\src_user\Settings\TlmCmd\common_cmd_packet_define.c"
-call :sync_file ".\src\src_user\Settings\TlmCmd\common_tlm_cmd_packet_define.h" "..\mobc\src\src_user\Settings\TlmCmd\common_tlm_cmd_packet_define.h"
-call :sync_file ".\src\src_user\TlmCmd\block_command_user_settings.c" "..\mobc\src\src_user\TlmCmd\block_command_user_settings.c"
-call :sync_file ".\src\src_user\TlmCmd\block_command_user_settings.h" "..\mobc\src\src_user\TlmCmd\block_command_user_settings.h"
-call :sync_file ".\src\src_user\TlmCmd\command_source.h" "..\mobc\src\src_user\TlmCmd\command_source.h"
-call :sync_file ".\src\src_user\TlmCmd\common_cmd_packet.c" "..\mobc\src\src_user\TlmCmd\common_cmd_packet.c"
-call :sync_file ".\src\src_user\TlmCmd\common_tlm_cmd_packet.c" "..\mobc\src\src_user\TlmCmd\common_tlm_cmd_packet.c"
+call :sync_file ".\src\src_user\Settings\tlm_cmd\common_cmd_packet_define.c" "..\mobc\src\src_user\Settings\tlm_cmd\common_cmd_packet_define.c"
+call :sync_file ".\src\src_user\Settings\tlm_cmd\common_tlm_cmd_packet_define.h" "..\mobc\src\src_user\Settings\tlm_cmd\common_tlm_cmd_packet_define.h"
+call :sync_file ".\src\src_user\tlm_cmd\block_command_user_settings.c" "..\mobc\src\src_user\tlm_cmd\block_command_user_settings.c"
+call :sync_file ".\src\src_user\tlm_cmd\block_command_user_settings.h" "..\mobc\src\src_user\tlm_cmd\block_command_user_settings.h"
+call :sync_file ".\src\src_user\tlm_cmd\command_source.h" "..\mobc\src\src_user\tlm_cmd\command_source.h"
+call :sync_file ".\src\src_user\tlm_cmd\common_cmd_packet.c" "..\mobc\src\src_user\tlm_cmd\common_cmd_packet.c"
+call :sync_file ".\src\src_user\tlm_cmd\common_tlm_cmd_packet.c" "..\mobc\src\src_user\tlm_cmd\common_tlm_cmd_packet.c"
 
 call :sync_file ".\src\src_user\Test\utils\wings_utils.py" "..\mobc\src\src_user\Test\utils\wings_utils.py"
 call :sync_file ".\src\src_user\Test\pytest.ini" "..\mobc\src\src_user\Test\pytest.ini"
